Code cleaning for clean install
