+++ indexedconv/__init__.py
@@ -1,1 +1,3 @@
-from . import engine
+import indexedconv.utils
+import indexedconv.engine
+import indexedconv.nets
